fix typos

+++ player.go
@@ -32,7 +32,7 @@ func newPlayer(renderer *sdl.Renderer) (player *element) {
 	mover := newKeyboardMover(player, playerSpeed)
 	player.addComponent(mover)
 
-	shooter := newKeyboardMover(player, playerShotCooldown)
+	shooter := newKeyboardMover(player, float64(playerShotCooldown))
 	player.addComponent((shooter))
 
 	return player
